Completed StartGame function logic
Created a separate "OneRound" function that simulates a round's loop. Also completed the "StartGame" function, which treats the cases in which the minimum number of cards that have to be dropped is either 1 or 2 (depending on if the deck is empty or not), and also allows the player to pick other cards to drop (if they have any more available).